refactor: extract formatHousing to utils
