+++ connectors/edf/src/utils.js
@@ -114,3 +114,79 @@ export function getEnergyTypeFromContract(contract) {
     ? 'electricity'
     : 'gas'
 }
+
+export function formatHousing(
+  contracts,
+  echeancierResult,
+  {
+    constructionDate,
+    equipment,
+    heatingSystem,
+    housingType,
+    lifeStyle,
+    surfaceInSqMeter,
+    residenceType,
+    contractElec,
+    rawConsumptions,
+  },
+) {
+  const consumptions = {
+    electricity: convertConsumption(
+      get(rawConsumptions, 'elec.yearlyElecEnergies'),
+      get(rawConsumptions, 'elec.monthlyElecEnergies'),
+    ),
+    gas: convertConsumption(
+      get(rawConsumptions, 'gas.yearlyGasEnergies'),
+      get(rawConsumptions, 'gas.monthlyGasEnergies'),
+    ),
+  }
+
+  const result = []
+  for (const key in contracts.details) {
+    const detail = contracts.details[key]
+    const energyProviders = detail.contracts.map(c => {
+      const energyType = getEnergyTypeFromContract(c)
+      const mappedContract = {
+        vendor: 'edfclientside',
+        contract_number: c.number,
+        energy_type: energyType,
+        contract_type: get(c, 'subscribeOffer.offerName'),
+        powerkVA: parseInt(
+          get(contractElec, 'supplyContractParameters.SUBSCRIBED_POWER'),
+          10,
+        ),
+        [energyType + '_consumptions']: consumptions[energyType],
+        charging_type: echeancierResult.isMonthly ? 'monthly' : 'yearly',
+      }
+
+      // even if the api does not show it, real pdl number for gas is pce_number
+      const pdlKeyMap = {
+        electricity: 'pdl_number',
+        gas: 'pce_number',
+      }
+      return {
+        ...mappedContract,
+        [pdlKeyMap[energyType]]: c.pdlnumber,
+      }
+    })
+    const housing = {
+      construction_year: constructionDate,
+      residence_type: convertResidenceType(residenceType),
+      housing_type: convertHousingType(housingType),
+      residents_number: lifeStyle.noOfOccupants,
+      living_space_m2: surfaceInSqMeter,
+      heating_system: convertHeatingSystem(
+        heatingSystem.principalHeatingSystemType,
+      ),
+      water_heating_system: convertWaterHeatingSystem(
+        get(equipment, 'sanitoryHotWater.sanitoryHotWaterType'),
+      ),
+      baking_types: convertBakingTypes(equipment.cookingEquipment),
+      address: detail.adress,
+      energy_providers: energyProviders,
+    }
+    result.push(housing)
+  }
+
+  return result
+}
